# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MBS CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/lead-distribution:
    get:
      operationId: getLeadDistributionSettings
      tags: [admin]
      summary: Get inbound lead distribution settings (admin only)
      responses:
        "200":
          description: Inbound lead distribution settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDistributionSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

    put:
      operationId: updateLeadDistributionSettings
      tags: [admin]
      summary: Update inbound lead distribution settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadDistributionSettingsUpdate"
      responses:
        "200":
          description: Updated inbound lead distribution settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDistributionSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  /flyers/generate:
    post:
      operationId: generateFlyer
      tags: [flyers]
      summary: Generate a PDF flyer from a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateFlyerInput"
      responses:
        "200":
          description: Generated flyer info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateFlyerResponse"
  /flyers/{id}/download:
    get:
      operationId: downloadFlyer
      tags: [flyers]
      summary: Download a generated flyer PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
  /flyers/{id}/email:
    post:
      operationId: emailFlyer
      tags: [flyers]
      summary: Email a generated flyer to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailFlyerInput"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /flyers/{id}:
    get:
      operationId: getGeneratedFlyer
      tags: [flyers]
      summary: Get a generated flyer record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated flyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedFlyer"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LeadDistributionSettings:
      type: object
      required: [includeAdminsInRoundRobin, staleThresholdDays]
      properties:
        includeAdminsInRoundRobin:
          type: boolean
          default: false
          description: Include active admins after active reps and managers in inbound round-robin assignment
        staleThresholdDays:
          type: integer
          minimum: 1
          maximum: 365
          default: 7
          description: Number of idle days before an assigned lead is considered stale
    LeadDistributionSettingsUpdate:
      type: object
      required: []
      properties:
        includeAdminsInRoundRobin:
          type: boolean
        staleThresholdDays:
          type: integer
          minimum: 1
          maximum: 365
...
    VariableField:
      type: object
      required: [key, label, type, defaultValue]
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, number, select]
        defaultValue:
          type: string
        options:
          type: array
          items:
            type: string
    FlyerTemplate:
      type: object
      required: [id, name, programType, htmlTemplate, variableFields, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        programType:
          type: string
          enum: [equipment, working_capital, general]
        htmlTemplate:
          type: string
        variableFields:
          type: array
          items:
            $ref: "#/components/schemas/VariableField"
        isActive:
          type: boolean
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    GeneratedFlyer:
      type: object
      required: [id, templateId, fieldValues, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: ["integer", "null"]
        templateId:
          type: ["integer", "null"]
        template:
          oneOf:
            - $ref: "#/components/schemas/FlyerTemplate"
            - type: "null"
        fieldValues:
          type: object
          additionalProperties:
            type: string
        pdfStorageKey:
          type: ["string", "null"]
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    GenerateFlyerInput:
      type: object
      required: [templateId, fieldValues]
      properties:
        templateId:
          type: integer
        fieldValues:
          type: object
          additionalProperties:
            type: string
        leadId:
          type: ["integer", "null"]
    GenerateFlyerResponse:
      type: object
      required: [flyerId]
      properties:
        flyerId:
          type: integer
        downloadUrl:
          type: string
    EmailFlyerInput:
      type: object
      required: [leadId]
      properties:
        leadId:
          type: integer
...